Another large file rename to be inline with dart conventions
