+++ example/lib/components/custom_text.dart
