+++ example/lib/main.dart
@@ -2,8 +2,8 @@ import 'package:flutter/material.dart';
 import 'package:rly_network_flutter_sdk/wallet.dart';
 
 import 'account_overview_screen.dart';
-import 'GenerateAccountScreen.dart';
-import 'LoadingScreen.dart';
+import 'generate_account_screen.dart';
+import 'loading_screen.dart';
 import 'package:rly_network_flutter_sdk/account.dart';
 
 class App extends StatefulWidget {
